Change documentables detection in update, refs #107
Before this change we used "filename" attribute to look for the
Primary object updated. The problem is that, that "filename" is,
in general:

- Not unique, there could be two different pod6 files with the same
name, but in different subdirectories.
- No canonical. For instance, the filename of a Type is X::Something,
instead of X/Something.pod6.

To avoid these problems we use an absolute path for each pod6 file,
`source-path`. `Pod::From::Cache` returns paths relatives to the working
directory of the cache, so they need to be converted to `absolute`.