Sort out situation with escaping URLs
And streamline forming of the search index entry
fragment URL.

Since the beginning, results of the url method for search index
entries and secondary pages were not percent encoded,
but instead did sort of escaping to work well with
the "static files on a UNIX file system" assumption.

The Index one also relied on various components of the X<>
formatting code, which is not desireable, instead it has to
refer only to exact text of the index entry itself, so that
other components, category and text, could be changed without
creating a 404 case.

This commit adds percent encoding to ensure:
- URLs we form for both anchors and secondary are understood
  properly by browsers, text editors and various software where
  a clickable URL is good
- We don't rely on various hacks to serve such URLs